Updated to the newest version of BotBase

+++ FBIBot/CommandHandler.cs
@@ -3,12 +3,12 @@ using Discord.Commands;
 using Discord.WebSocket;
 using FBIBot.Modules.AutoMod;
 using FBIBot.Modules.AutoMod.AutoSurveillance;
-using FBIBot.Modules.Config;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
+using static FBIBot.DatabaseManager;
 
 namespace FBIBot
 {
@@ -17,33 +17,33 @@ namespace FBIBot
         public const string prefix = "\\";
         public static int argPos = 0;
 
-        private readonly DiscordSocketClient _client;
-        private readonly CommandService _commands;
-        private readonly IServiceProvider _services;
+        private readonly DiscordSocketClient client;
+        private readonly CommandService commands;
+        private readonly IServiceProvider services;
 
         public CommandHandler(DiscordSocketClient client, IServiceProvider services)
         {
-            _client = client;
-            _services = services;
+            this.client = client;
+            this.services = services;
 
             CommandServiceConfig config = new CommandServiceConfig()
             {
                 DefaultRunMode = RunMode.Async
             };
-            _commands = new CommandService(config);
+            commands = new CommandService(config);
         }
 
         public async Task InitCommandsAsync()
         {
-            _client.Connected += SendConnectMessage;
-            _client.Disconnected += SendDisconnectError;
-            _client.JoinedGuild += SendJoinMessage;
-            _client.UserJoined += SendWelcomeMessage;
-            _client.GuildMemberUpdated += CheckUsernameAsync;
-            _client.MessageReceived += HandleCommandAsync;
+            client.Connected += SendConnectMessage;
+            client.Disconnected += SendDisconnectError;
+            client.JoinedGuild += SendJoinMessage;
+            client.UserJoined += SendWelcomeMessage;
+            client.GuildMemberUpdated += CheckUsernameAsync;
+            client.MessageReceived += HandleCommandAsync;
 
-            await _commands.AddModulesAsync(Assembly.GetEntryAssembly(), _services);
-            _commands.CommandExecuted += SendErrorAsync;
+            await commands.AddModulesAsync(Assembly.GetEntryAssembly(), services);
+            commands.CommandExecuted += SendErrorAsync;
         }
 
         private async Task SendErrorAsync(Optional<CommandInfo> info, ICommandContext context, IResult result)
@@ -56,30 +56,24 @@ namespace FBIBot
 
         private async Task SendConnectMessage()
         {
-            if (Program.isConsole)
-            {
-                await Console.Out.WriteLineAsync($"{SecurityInfo.botName} is online");
-            }
+            await Console.Out.WriteLineAsync($"{SecurityInfo.botName} is online");
         }
 
         private async Task SendDisconnectError(Exception e)
         {
-            if (Program.isConsole)
-            {
-                await Console.Out.WriteLineAsync(e.Message);
-            }
+            await Console.Out.WriteLineAsync(e.Message);
         }
 
         private async Task SendJoinMessage(SocketGuild g) => await g.DefaultChannel.SendMessageAsync("Someone called for some democracy and justice?");
 
         private async Task SendWelcomeMessage(SocketGuildUser u)
         {
-            if (await RaidMode.GetVerificationLevelAsync(u.Guild) != null && !u.IsBot)
+            if (await raidModeDatabase.RaidMode.GetVerificationLevelAsync(u.Guild) != null && !u.IsBot)
             {
                 await Task.WhenAll
                 (
                     u.SendMessageAsync($":rotating_light: :rotating_light: The FBI of {u.Guild.Name} is currently in Raid Mode. As a result, you may not join the server at this time.:rotating_light: :rotating_light:"),
-                    RaidMode.AddBlockedUserAsync(u)
+                    raidModeDatabase.UsersBlocked.AddBlockedUserAsync(u)
                 );
                 await u.KickAsync("FBI RAID MODE");
                 return;
@@ -100,15 +94,15 @@ namespace FBIBot
             int index = Program.rng.Next(messages.Count);
             await channel.SendMessageAsync($"{u.Mention} {messages[index]}");
 
-            if (await Verify.GetVerifiedAsync(u))
+            if (await verificationDatabase.Verified.GetVerifiedAsync(u))
             {
-                SocketRole role = await SetVerify.GetVerificationRoleAsync(u.Guild);
+                SocketRole role = await verificationDatabase.Roles.GetVerificationRoleAsync(u.Guild);
                 if (role != null && u.Guild.CurrentUser.GetPermissions(u.Guild.DefaultChannel).ManageRoles)
                 {
                     await u.AddRoleAsync(role);
                 }
             }
-            else if (!u.IsBot && await SetVerify.GetVerificationRoleAsync(u.Guild) != null)
+            else if (!u.IsBot && await verificationDatabase.Roles.GetVerificationRoleAsync(u.Guild) != null)
             {
                 await Verify.SendCaptchaAsync(u);
             }
@@ -117,7 +111,7 @@ namespace FBIBot
             Task<bool>[] isZalgo =
             {
                 Zalgo.IsZalgoAsync(name),
-                AntiZalgo.GetAntiZalgoAsync(u.Guild)
+                configDatabase.AntiZalgo.GetAntiZalgoAsync(u.Guild)
             };
 
             if ((await Task.WhenAll(isZalgo)).All(x => x))
@@ -138,7 +132,7 @@ namespace FBIBot
             Task<bool>[] isZalgo =
             {
                 Zalgo.IsZalgoAsync(newName),
-                AntiZalgo.GetAntiZalgoAsync(updated.Guild)
+                configDatabase.AntiZalgo.GetAntiZalgoAsync(updated.Guild)
             };
 
             if ((await Task.WhenAll(isZalgo)).All(x => x))
@@ -150,24 +144,26 @@ namespace FBIBot
                 catch { }
             }
         }
+        private async Task<bool> CanBotRunCommandsAsync(SocketUserMessage msg) => await Task.Run(() => msg.Author.Id == client.CurrentUser.Id);
+        private async Task<bool> ShouldDeleteBotCommands() => await Task.Run(() => true);
 
         private async Task HandleCommandAsync(SocketMessage m)
         {
-            if (!(m is SocketUserMessage msg) || (msg.Author.IsBot && msg.Author.Id != _client.CurrentUser.Id))
+            if (!(m is SocketUserMessage msg) || (msg.Author.IsBot && !await CanBotRunCommandsAsync(msg)))
             {
                 return;
             }
 
-            SocketCommandContext Context = new SocketCommandContext(_client, msg);
-            string _prefix = Context.Guild != null ? await SetPrefix.GetPrefixAsync(Context.Guild) : prefix;
-            bool isCommand = msg.HasMentionPrefix(_client.CurrentUser, ref argPos) || msg.HasStringPrefix(_prefix, ref argPos);
+            SocketCommandContext Context = new SocketCommandContext(client, msg);
+            string _prefix = Context.Guild != null ? await configDatabase.Prefixes.GetPrefixAsync(Context.Guild) : prefix;
+            bool isCommand = msg.HasMentionPrefix(client.CurrentUser, ref argPos) || msg.HasStringPrefix(_prefix, ref argPos);
 
             if (isCommand)
             {
-                var result = await _commands.ExecuteAsync(Context, argPos, _services);
+                var result = await commands.ExecuteAsync(Context, argPos, services);
 
                 List<Task> cmds = new List<Task>();
-                if (msg.Author.IsBot)
+                if (msg.Author.IsBot && await ShouldDeleteBotCommands())
                 {
                     cmds.Add(msg.DeleteAsync());
                 }
@@ -190,40 +186,40 @@ namespace FBIBot
             Task<bool>[] isZalgo =
             {
                 Zalgo.IsZalgoAsync(Context),
-                AntiZalgo.GetAntiZalgoAsync(Context.Guild)
+                configDatabase.AntiZalgo.GetAntiZalgoAsync(Context.Guild)
             };
             Task<bool>[] isSpam =
             {
                 Spam.IsSpamAsync(Context),
-                AntiSpam.GetAntiSpamAsync(Context.Guild)
+                configDatabase.AntiSpam.GetAntiSpamAsync(Context.Guild)
             };
             Task<bool>[] isSingleSpam =
             {
                 Spam.IsSingleSpamAsync(Context),
-                AntiSingleSpam.GetAntiSingleSpamAsync(Context.Guild)
+                configDatabase.AntiSingleSpam.GetAntiSingleSpamAsync(Context.Guild)
             };
             Task<bool>[] isMassMention =
             {
                 MassMention.IsMassMentionAsync(Context),
-                AntiMassMention.GetAntiMassMentionAsync(Context.Guild)
+                configDatabase.AntiMassMention.GetAntiMassMentionAsync(Context.Guild)
             };
             Task<bool>[] isCaps =
             {
                 CAPS.ISCAPSASYNC(Context),
-                AntiCaps.GetAntiCapsAsync(Context.Guild)
+                configDatabase.AntiCaps.GetAntiCapsAsync(Context.Guild)
             };
             Task<bool>[] isInvite =
             {
                 Invite.HasInviteAsync(Context),
-                AntiInvite.GetAntiInviteAsync(Context.Guild)
+                configDatabase.AntiInvite.GetAntiInviteAsync(Context.Guild)
             };
             Task<bool>[] isLink =
             {
                 Link.IsLinkAsync(Context),
-                AntiLink.GetAntiLinkAsync(Context.Guild)
+                configDatabase.AntiLink.GetAntiLinkAsync(Context.Guild)
             };
 
-            if (await AutoSurveillance.GetAutoSurveillanceAsync(Context.Guild))
+            if (await configDatabase.AutoSurveillance.GetAutoSurveillanceAsync(Context.Guild))
             {
                 if (await AutoSurveillanceAsync(Context))
                 {
